pyproject.toml
[project]
name = "my-project"
version = "0.1.0"
description = "Production-ready multi-agent sales analytics assistant."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai>=1.3.0",
    "langchain>=1.0.3",
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "openai>=1.109.1",
    "pandas>=2.3.3",
    "scikit-learn>=1.7.2",
    "streamlit>=1.51.0",
    "litellm>=1.52.0",
    "langchain-groq>=1.0.0",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["setuptools>=67.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"my_project" = ["config/*.yaml", "data/*.csv"]
